Fix data race issue on copyFn when there are multiple containers or initContainers (#422)
Fixes #421 

Creates a new `ImageCopier` object with its own `imagePullPolicy` to
avoid a data race on the `container` object.

This new object is dedicated to the image copy process as it was done
before but with its own subcontext and abstracted from the core mutation
webhook processing.

A new configurable `imageCopyDeadline` parameter has been added
alongside an implementation of timeout throughout the different steps of
the image copy. Every function called as part of the copy has been
updated to include context evaluation and support abortion in case of
context timeout.